chore: split apply_patch logic out of codex.rs and into apply_patch.rs
